pyproject.toml
[project]
name = "los-classificados"
version = "0.1.0"
description = "Digital marketplace connecting users with local service providers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=4.1.0",
    "dash-bootstrap-components>=2.0.0",
    "plotly>=5.22.0",
    "flask>=3.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "dash-iconify>=0.1.2",
    "pandas>=2.2.0",
    "cryptography>=42.0.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
